refactor: move ATen e2e runner out of generator
